+++ Tests/SentryTests/SentrySessionTests.m
@@ -9,24 +9,24 @@
 
 - (void)testInitDefaultValues
 {
-    SentrySession *session = [[SentrySession alloc] init];
+    SentrySession *session = [[SentrySession alloc] initWithReleaseName:@"1.0.0"];
     XCTAssertNotNil(session.sessionId);
     XCTAssertEqual(1, session.sequence);
     XCTAssertEqual(0, session.errors);
-    XCTAssertTrue(session.init);
+    XCTAssertTrue(session.flagInit);
     XCTAssertNotNil(session.started);
     XCTAssertEqual(kSentrySessionStatusOk, session.status);
     XCTAssertNotNil(session.distinctId);
 
     XCTAssertNil(session.timestamp);
-    XCTAssertNil(session.releaseName);
+    XCTAssertEqual(@"1.0.0", session.releaseName);
     XCTAssertNil(session.environment);
     XCTAssertNil(session.duration);
 }
 
 - (void)testSerializeDefaultValues
 {
-    SentrySession *expected = [[SentrySession alloc] init];
+    SentrySession *expected = [[SentrySession alloc] initWithReleaseName:@"1.0.0"];
     NSDictionary<NSString *, id> *json = [expected serialize];
     SentrySession *actual = [[SentrySession alloc] initWithJSONObject:json];
 
@@ -41,8 +41,8 @@
     XCTAssertNil(expected.timestamp);
     // Serialize session always have a timestamp (time of serialization)
     XCTAssertNotNil(actual.timestamp);
-    XCTAssertNil(expected.releaseName);
-    XCTAssertNil(actual.releaseName);
+    XCTAssertEqual(@"1.0.0", expected.releaseName);
+    XCTAssertEqual(@"1.0.0", actual.releaseName);
     XCTAssertNil(expected.environment);
     XCTAssertNil(actual.environment);
     XCTAssertNil(expected.duration);
@@ -51,11 +51,10 @@
 
 - (void)testSerializeExtraFieldsEndedSessionWithNilStatus
 {
-    SentrySession *expected = [[SentrySession alloc] init];
+    SentrySession *expected = [[SentrySession alloc] initWithReleaseName:@"io.sentry@5.0.0-test"];
     NSDate *timestamp = [NSDate date];
     [expected endSessionExitedWithTimestamp:timestamp];
     expected.environment = @"prod";
-    expected.releaseName = @"io.sentry@5.0.0-test";
     NSDictionary<NSString *, id> *json = [expected serialize];
     SentrySession *actual = [[SentrySession alloc] initWithJSONObject:json];
 
@@ -78,7 +77,7 @@
 
 - (void)testSerializeErrorIncremented
 {
-    SentrySession *expected = [[SentrySession alloc] init];
+    SentrySession *expected = [[SentrySession alloc] initWithReleaseName:@""];
     [expected incrementErrors];
     [expected endSessionExitedWithTimestamp:[NSDate date]];
     NSDictionary<NSString *, id> *json = [expected serialize];
@@ -101,7 +100,7 @@
 
 - (void)testAbnormalSession
 {
-    SentrySession *expected = [[SentrySession alloc] init];
+    SentrySession *expected = [[SentrySession alloc] initWithReleaseName:@""];
     XCTAssertEqual(0, expected.errors);
     XCTAssertEqual(kSentrySessionStatusOk, expected.status);
     XCTAssertEqual(1, expected.sequence);
@@ -117,7 +116,7 @@
 
 - (void)testCrashedSession
 {
-    SentrySession *expected = [[SentrySession alloc] init];
+    SentrySession *expected = [[SentrySession alloc] initWithReleaseName:@""];
     XCTAssertEqual(1, expected.sequence);
     XCTAssertEqual(kSentrySessionStatusOk, expected.status);
     [expected endSessionCrashedWithTimestamp:[NSDate date]];
@@ -127,7 +126,7 @@
 
 - (void)testExitedSession
 {
-    SentrySession *expected = [[SentrySession alloc] init];
+    SentrySession *expected = [[SentrySession alloc] initWithReleaseName:@""];
     XCTAssertEqual(0, expected.errors);
     XCTAssertEqual(kSentrySessionStatusOk, expected.status);
     XCTAssertEqual(1, expected.sequence);